+++ killstats/api/helpers.py
@@ -79,15 +79,16 @@ def get_main_and_alts_all(corporations: list):
     for member in corpmember.objects.filter(
         corpstats__corp__corporation_id__in=corporations
     ).exclude(character_id__in=chars_list):
-        char = (
-            EveCharacter.objects.select_related(
+        try:
+            char = EveCharacter.objects.select_related(
                 "character_ownership",
                 "character_ownership__user__profile__main_character",
+            ).get(character_id=member.character_id)
+            _process_character(
+                char, characters, chars_list, corporations, missing_chars
             )
-            .prefetch_related("character_ownership__user__character_ownerships")
-            .get(character_id=member.character_id)
-        )
-        _process_character(char, characters, chars_list, corporations, missing_chars)
+        except ObjectDoesNotExist:
+            missing_chars.add(member.character_id)
 
     # TODO Maybe create task for missing_chars if needed
 
@@ -95,9 +96,11 @@ def get_main_and_alts_all(corporations: list):
 
 
 def get_corporations(request):
-    linked_characters = request.user.profile.main_character.character_ownership.user.character_ownerships.all().values_list(
-        "character_id", flat=True
-    )
+    linked_characters = request.user.profile.main_character.character_ownership.user.character_ownerships.select_related(
+        "character", "user"
+    ).all()
+
+    linked_characters = linked_characters.values_list("character_id", flat=True)
     chars = EveCharacter.objects.filter(id__in=linked_characters)
 
     corporations = set()
@@ -105,4 +108,4 @@ def get_corporations(request):
     for char in chars:
         corporations.add(char.corporation_id)
 
-    return corporations
+    return list(corporations)